extract getFields to a new class
